% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{predict10}
\alias{predict10}
\title{prediction with the top 10 features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{a n x 10 matrix containing n new observations}

\item{b}{a vector of coefficients obtained from the \code{topten} function}
}
\value{
a numeric vector containing the predictor values
}
\description{
This function takes a set of coefficients produced by the \code{topten}
and makes a prediction for each of the values provided in the input X
matrix
}
